Delegadas as funções de verificação e exceção para a camada de serviço do Seller.

+++ Application/Services/SellerService.cs
@@ -2,26 +2,56 @@
 using Domain.Interfaces.Application.Services;
 using Domain.Interfaces.Infra.Repositories;
 using Domain.Models;
+using Domain.ViewModels;
 
 namespace Application.Services
 {
     public class SellerService : ISellerService
     {
         private readonly ISellerRepository _sellerRepository;
+        private readonly IDepartmentRepository _departmentRepository;
 
-        public SellerService(ISellerRepository sellerRepository)
+        public SellerService(ISellerRepository sellerRepository, IDepartmentRepository departmentRepository)
         {
             _sellerRepository = sellerRepository;
+            _departmentRepository = departmentRepository;
         }
 
         public OperationResult<Seller> Create(Seller seller)
         {
-            return _sellerRepository.Create(seller);
+            try
+            {
+                var result = _sellerRepository.Create(seller);
+                if (result == null)
+                    return OperationResult<Seller>.CreateFail("Seller not found.");
+
+                return OperationResult<Seller>.CreateSuccess(seller);
+            }
+            catch
+            {
+                return OperationResult<Seller>.CreateFail("There was an unexpected error, please try again.");
+            }
         }
 
         public OperationResult Delete(int id)
         {
-            return _sellerRepository.Delete(id);
+            try
+            {
+                if (id == 0)
+                    return OperationResult.CreateFail("Id not found.");
+
+                var result = _sellerRepository.GetById(id);
+                if (result == null)
+                    return OperationResult.CreateFail("Seller not found.");
+
+                _sellerRepository.Delete(id);
+
+                return OperationResult.CreateSuccess();
+            }
+            catch
+            {
+                return OperationResult.CreateFail("There was an unexpected error, please try again.");
+            }
         }
 
         public List<Seller> FindAll()
@@ -29,14 +59,42 @@ namespace Application.Services
             return _sellerRepository.FindAll();
         }
 
-        public Seller GetById(int id)
+        public OperationResult<SellerFormViewModel> GetById(int id)
         {
-            return _sellerRepository.GetById(id);
+            try
+            {
+                if (id == 0)
+                    return OperationResult<SellerFormViewModel>.CreateFail("Id not found");
+
+                var obj = _sellerRepository.GetById(id);
+                if (obj == null)
+                    return OperationResult<SellerFormViewModel>.CreateFail("Seller not found");
+
+                var departments = _departmentRepository.FindAll();
+                var sellerForm = new SellerFormViewModel { Departments = departments, Seller = obj };
+
+                return OperationResult<SellerFormViewModel>.CreateSuccess(sellerForm);
+            }
+            catch
+            {
+                return OperationResult<SellerFormViewModel>.CreateFail("There was an unexpected error, please try again");
+            }
         }
 
         public OperationResult<Seller> Edit(Seller seller)
         {
-            return _sellerRepository.Edit(seller);
-        }
+            try
+            {
+                var result = _sellerRepository.Edit(seller);
+                if (result == null)
+                    return OperationResult<Seller>.CreateFail("Seller not found");
+
+                return OperationResult<Seller>.CreateSuccess(seller);
+            }
+            catch
+            {
+                return OperationResult<Seller>.CreateFail("There was an unexpected error, please try again");
+            }
+        } 
     }
 }
